requirements.txt
asgiref==3.8.1
Django==5.1.6
django-cors-headers==4.7.0
django-filter==25.1
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
django-filter==25.1
pillow==11.1.0
pip==24.3.1
PyJWT==2.9.0
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.1